Database integration
Doesn't work properly yet.
User and Template can be saved to database properly.
User can be retrieved from database. An issue with Group is causing an error to occur when trying to retrieve Template.
An issue with Date is preventing Group to be saved to the database properly. The database throws a MySQL error.